requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.2.0
annotated-types==0.6.0
anyio==4.6.2
asttokens==3.0.0
attrs==24.3.0
Brotli==1.0.9
certifi==2025.1.31
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.13
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.2.2
executing==2.1.0
faiss==1.9.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
idna==3.7
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.6.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.2
multidict==6.1.0
nest_asyncio==1.6.0
numpy==1.26.4
openai==0.28.1
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pip==25.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
pyzmq==26.3.0
requests==2.32.3
setuptools==72.1.0
six==1.17.0
sniffio==1.3.0
stack_data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
yarl==1.18.0
zipp==3.21.0
